# typespec/tsp-output/@typespec/openapi3/openapi-1.0.0.yaml
openapi: 3.0.0
info:
  title: Julep API
  termsOfService: https://julep.ai/terms
  contact:
    name: Julep AI
    url: https://julep.ai
    email: developers@julep.ai
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  summary: A backend for creating stateful AI apps
  description: Julep is a backend for creating stateful AI apps with background tasks and long-term memory easily.
  version: 1.0.0
...
paths:
...
  /agents/{id}/docs:
    get:
      operationId: AgentDocsRoute_list
      description: List Docs owned by an Agent
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent
          schema:
            $ref: '#/components/schemas/Common.uuid'
        - $ref: '#/components/parameters/Common.PaginationOptions.limit'
        - $ref: '#/components/parameters/Common.PaginationOptions.offset'
        - $ref: '#/components/parameters/Common.PaginationOptions.sort_by'
        - $ref: '#/components/parameters/Common.PaginationOptions.direction'
        - $ref: '#/components/parameters/Common.PaginationOptions.metadata_filter'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Docs.Doc'
                required:
                  - items
    post:
      operationId: AgentDocsRoute_create
      description: Create a Doc for this Agent
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent resource
          schema:
            $ref: '#/components/schemas/Common.uuid'
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Common.ResourceCreatedResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Docs.CreateDocRequest'
...
  /users/{id}/docs:
    get:
      operationId: UserDocsRoute_list
      description: List Docs owned by a User
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent
          schema:
            $ref: '#/components/schemas/Common.uuid'
        - $ref: '#/components/parameters/Common.PaginationOptions.limit'
        - $ref: '#/components/parameters/Common.PaginationOptions.offset'
        - $ref: '#/components/parameters/Common.PaginationOptions.sort_by'
        - $ref: '#/components/parameters/Common.PaginationOptions.direction'
        - $ref: '#/components/parameters/Common.PaginationOptions.metadata_filter'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Docs.Doc'
                required:
                  - items
    post:
      operationId: UserDocsRoute_create
      description: Create a Doc for this User
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent resource
          schema:
            $ref: '#/components/schemas/Common.uuid'
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Common.ResourceCreatedResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Docs.CreateDocRequest'
...
security:
  - ApiKeyAuth: []
  - ApiKeyAuth_: []
components:
  parameters:
...
    Common.PaginationOptions.direction:
      name: direction
      in: query
      required: true
      description: Sort direction
      schema:
        type: string
        enum:
          - asc
          - desc
        default: asc
      explode: false
    Common.PaginationOptions.limit:
      name: limit
      in: query
      required: true
      description: Limit the number of items returned
      schema:
        $ref: '#/components/schemas/Common.limit'
        default: 100
      explode: false
    Common.PaginationOptions.metadata_filter:
      name: metadata_filter
      in: query
      required: true
      description: Object to filter results by metadata
      schema:
        type: object
        additionalProperties:
          anyOf:
            - type: number
            - type: string
            - type: boolean
          nullable: true
      explode: false
    Common.PaginationOptions.offset:
      name: offset
      in: query
      required: true
      description: Offset the items returned
      schema:
        $ref: '#/components/schemas/Common.offset'
        default: 0
      explode: false
    Common.PaginationOptions.sort_by:
      name: sort_by
      in: query
      required: true
      description: Sort by a field
      schema:
        type: string
        enum:
          - created_at
          - updated_at
        default: created_at
      explode: false
...
  schemas:
...
    Common.ResourceCreatedResponse:
      type: object
      required:
        - id
        - created_at
        - jobs
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Common.uuid'
          description: ID of created resource
        created_at:
          type: string
          format: date-time
          description: When this resource was created as UTC date-time
          readOnly: true
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Common.uuid'
          description: IDs (if any) of jobs created as part of this request
          default: []
          readOnly: true
...
    Common.limit:
      type: integer
      format: uint16
      minimum: 1
      maximum: 1000
      exclusiveMaximum: true
      description: Limit the number of results
...
    Common.offset:
      type: integer
      format: uint32
      minimum: 0
      description: Offset to apply to the results
    Common.uuid:
      type: string
      format: uuid
...
    Docs.CreateDocRequest:
      type: object
      required:
        - title
        - content
        - embed_instruction
      properties:
        metadata:
          type: object
          additionalProperties: {}
        title:
          type: string
          maxLength: 800
          description: Title describing what this document contains
        content:
          anyOf:
            - type: string
            - type: array
              items:
                type: string
          description: Contents of the document
        embeddings:
          anyOf:
            - type: array
              items:
                type: number
                format: float
            - type: array
              items:
                type: array
                items:
                  type: number
                  format: float
          description: Embeddings for the document
        embed_instruction:
          type: string
          nullable: true
          description: Instruction for the embedding model.
          default: null
      description: Payload for creating a doc
    Docs.Doc:
      type: object
      required:
        - id
        - created_at
        - title
        - content
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Common.uuid'
          readOnly: true
        metadata:
          type: object
          additionalProperties: {}
        created_at:
          type: string
          format: date-time
          description: When this resource was created as UTC date-time
          readOnly: true
        title:
          type: string
          maxLength: 800
          description: Title describing what this document contains
        content:
          anyOf:
            - type: string
            - type: array
              items:
                type: string
          description: Contents of the document
        embeddings:
          anyOf:
            - type: array
              items:
                type: number
                format: float
            - type: array
              items:
                type: array
                items:
                  type: number
                  format: float
          description: Embeddings for the document
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    ApiKeyAuth_:
      type: apiKey
      in: header
      name: X-Auth-Key
servers:
  - url: https://{serverEnv}.julep.ai/api
    description: The julep cloud service endpoint
    variables:
      serverEnv:
        default: api-alpha
        description: The environment to use
        enum:
          - api
          - api-alpha